+++ packages/elements/code/examples/00-inline-code-basic.js
@@ -7,8 +7,7 @@ const exampleInlineCode = `const map = new Map({ key: 'value' })`;
 export default function Component() {
   return (
     <span>
-      This is inline code:{' '}
-      <AkCode language="javascript" text={exampleInlineCode} />, check it out.
+      This is inline code: <AkCode text={exampleInlineCode} />, check it out.
     </span>
   );
 }