+++ packages/elements/input/src/index.js
@@ -1,0 +1,4 @@
+// @flow
+import SingleLineTextInput from './SingleLineTextInput';
+
+export default SingleLineTextInput;