+++ packages/fabric/editor-core/src/nodeviews/ui/panel.tsx
@@ -10,35 +10,41 @@ import {
   akColorP50,
   akColorB50,
   akColorY50,
+  akColorR50,
   akColorG400,
   akColorP400,
   akColorB400,
   akColorY400,
+  akColorR400,
 } from '@atlaskit/util-shared-styles';
 import InfoIcon from '@atlaskit/icon/glyph/editor/info';
-import TipIcon from '@atlaskit/icon/glyph/editor/hint';
+import SuccessIcon from '@atlaskit/icon/glyph/editor/success';
 import NoteIcon from '@atlaskit/icon/glyph/editor/note';
-import WarningIcon from '@atlaskit/icon/glyph/warning';
+import WarningIcon from '@atlaskit/icon/glyph/editor/warning';
+import ErrorIcon from '@atlaskit/icon/glyph/editor/error';
 
 const panelColor = {
   info: akColorB50,
   note: akColorP50,
-  tip: akColorG50,
+  success: akColorG50,
   warning: akColorY50,
+  error: akColorR50,
 };
 
 const iconColor = {
   info: akColorB400,
   note: akColorP400,
-  tip: akColorG400,
+  success: akColorG400,
   warning: akColorY400,
+  error: akColorR400,
 };
 
 const panelIcons = {
   info: InfoIcon,
-  tip: TipIcon,
+  success: SuccessIcon,
   note: NoteIcon,
   warning: WarningIcon,
+  error: ErrorIcon,
 };
 
 // tslint:disable-next-line:variable-name
@@ -72,8 +78,10 @@ class Panel implements NodeView {
   private domRef: HTMLElement | undefined;
   private contentDOMRef: HTMLElement | undefined;
   private panelType: string;
+  private node: PMNode;
 
   constructor(node: PMNode, view: EditorView, getPos: getPosHandler) {
+    this.node = node;
     this.panelType = node.attrs.panelType;
     this.renderReactComponent();
   }
@@ -109,12 +117,14 @@ class Panel implements NodeView {
     return this.contentDOMRef;
   }
 
-  update() {
-    /**
-     * Returning false here fixes an error where the editor fails to set selection
-     * inside the contentDOM after a transaction. See ED-2374.
-     */
-    return false;
+  update(node) {
+    if (
+      node.attrs.panelType !== this.node.attrs.panelType ||
+      node.type !== this.node.type
+    ) {
+      return false;
+    }
+    return true;
   }
 
   destroy() {