+++ packages/fabric/mention/__tests__/IndexSpec.tsx
@@ -1,0 +1,10 @@
+import Picker, * as other from '../src';
+
+describe(name, () => {
+  describe('exports', () => {
+    it('should export a base component', () => {
+      expect(Picker).not.toEqual(null);
+      expect(other.default).toEqual(Picker);
+    });
+  });
+});