+++ packages/fabric/mention/src/components/Scrollable/styles.ts
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+import { akBorderRadius } from '@atlaskit/util-shared-styles';
+import { scrollableMaxHeight } from '../../shared-styles';
+
+// tslint:disable:next-line variable-name
+export const ScrollableStyle = styled.div`
+  display: block;
+  overflowx: hidden;
+  overflowy: auto;
+
+  padding: 4px 0;
+  margin: 0;
+
+  background: white;
+  maxheight: ${scrollableMaxHeight};
+
+  borderradius: ${akBorderRadius};
+`;