+++ packages/elements/spinner/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/spinner
 
+## 4.1.2
+- [patch] Fix spinner performance by statically defining keyframes [6f04599](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6f04599)
+
 ## 4.1.1
 - [patch] Resolved low hanging flow errors in field-base field-text comment icon item and website, $ [007de27](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/007de27)