+++ packages/fabric/editor-confluence-transformer/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-cq-transformer
 
+## 2.7.0
+- [minor] added new panelType [9f693b9](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9f693b9)
+
+## 2.6.1
+
 ## 2.6.0
 - [minor] added date plugin [f7b8a33](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f7b8a33)