+++ packages/fabric/editor-json-transformer/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-json-transformer
 
+## 2.5.1
+
+## 2.5.0
+- [minor] Implement JSONTransformer::parse to enable ADF -> PM conversion [c8c3c9e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c8c3c9e)
+
 ## 2.4.0
 - [minor] Add React 16 support. [12ea6e4](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/12ea6e4)