+++ packages/fabric/media-filmstrip/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/media-filmstrip
 
+## 6.0.0
+
+## 5.8.1
+
 ## 5.8.0
 - [minor] Add React 16 support. [12ea6e4](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/12ea6e4)